renamed decrypt and encrypt classes to be decrypt_runner and encrypt_runner
