+++ spec/decrypt_spec.rb
@@ -1,19 +1,19 @@
-require './lib/decrypt'
+require './lib/decrypt_runner'
 require './lib/enigma'
 
-describe Decrypt do
+describe DecryptRunner do
   before :each do
-    @decrypt = Decrypt.new("encrypted.txt", "decrypted.txt", 82648, 240818)
+    @decrypt_runner = DecryptRunner.new("encrypted.txt", "decrypted.txt", 82648, 240818)
   end
 
   it 'exists' do
-    expect(@decrypt).to be_a(DecryptRunner)
+    expect(@decrypt_runner).to be_a(DecryptRunner)
   end
 
   it 'has a read file path, write file path, key, and date' do
-    expect(@decrypt.read_file_path).to eq("encrypted.txt")
-    expect(@decrypt.write_file_path).to eq("decrypted.txt")
-    expect(@decrypt.key).to eq(82648)
-    expect(@decrypt.date).to eq(240818)
+    expect(@decrypt_runner.read_file_path).to eq("encrypted.txt")
+    expect(@decrypt_runner.write_file_path).to eq("decrypted.txt")
+    expect(@decrypt_runner.key).to eq(82648)
+    expect(@decrypt_runner.date).to eq(240818)
   end
 end
